finally have client working pdf download
dumb problems with blank pdf output
