+++ .env.example
@@ -12,3 +12,7 @@ LOG_PATH=
 CERT_PEM_FILE_PATH=
 PRIV_KEY_FILE_PATH=
 CIPHERS_FILE_PATH=
+APP_BASE_PATH=
+LOCAL_APP_BASE_PATH=
+API_BASE_PATH=
+LOCAL_API_BASE_PATH=
