+++ index.js
@@ -43,8 +43,25 @@ if (process.env.NODE_ENV === "live") {
 
 // CORs
 app.use((req, res, next) => {
-    res.header("Access-Control-Allow-Origin", "*");
-    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    const allowedOriginsList = [
+        process.env.LOCAL_APP_BASE_PATH,
+        process.env.APP_BASE_PATH,
+    ];
+    const inboundOrigin = req.get('origin');
+    let allowedOrigin = '';
+
+    if (allowedOriginsList.indexOf(inboundOrigin) !== -1) {
+        allowedOrigin = inboundOrigin;
+    } else {
+        // not sure what the right status to send is here, this is probably not the right way to terminate CORs anyway
+        // but an improvemnet over wildcard
+        res.status(500).send('Bad CORS origin');
+        return;
+    }
+
+    res.header("Access-Control-Allow-Origin", allowedOrigin);
+    res.header('Access-Control-Allow-Methods', 'GET, POST');
+    res.header("Access-Control-Allow-Headers", "x-access-token, Origin, X-Requested-With, Content-Type, Accept");
     next();
 });
 
@@ -72,10 +89,8 @@ app.post('/login-user', loginUser);
 app.post('/upload-tag', verifyToken, uploadTags);
 app.post('/sync-up', verifyToken, syncUp); // these names are terrible
 app.post('/sync-down', verifyToken, syncDown);
-app.get('/generate-spreadsheet', verifyToken, generateSpreadsheet); // prod
-// app.get('/generate-spreadsheet', generateSpreadsheet); // dev
-// app.get('/generate-pdf', verifyToken, generatePdf); // prod
-app.get('/generate-pdf', generatePdf);
+app.post('/generate-spreadsheet', generateSpreadsheet); // no auth, partially due to difficulty (binary download) but also have to be logged in to trigger, files delete after download and filenames are obfuscated
+app.get('/generate-pdf*', generatePdf);
 
 if (process.env.NODE_ENV === "live") {
     https.createServer(https_options, app).listen(443);
