+++ utils/users/userFunctions.js
@@ -5,7 +5,7 @@ const bcrypt = require('bcrypt');
 const saltRounds = 15;
 
 // internal method currently, need middleware if public
-const createUser = (username, password) => {
+const _createUser = (username, password) => {
     if (!username || !password) {
         return false;
     }
@@ -48,7 +48,7 @@ const createUser = (username, password) => {
 }
 
 // internal method currently, need middleware if public
-const deleteUser = (userId) => {
+const _deleteUser = (userId) => {
     pool.query(
         `DELETE FROM users WHERE id = ?`,
         [userId],
@@ -90,6 +90,48 @@ const getUserIdFromToken = async (token) => {
     });
 }
 
+// I suppose it is possible to steal a sync_id on accident eg. race condition but it doesn't really matter
+// since it's just a unique reference
+const getSyncId = async (userId) => {
+    return new Promise(resolve => {
+        pool.query(
+            `INSERT INTO sync_history SET user_id = ?, sync_timestamp = ?`,
+            [userId, getDateTime()], // no sync id on uploads
+            (err, res) => {
+                if (err) {
+                    console.log('getSyncId', err);
+                    resolve(false);
+                } else {
+                    resolve(res.insertId);
+                }
+            }
+        );
+    });
+}
+
+const getRecentSyncId = (userId) => {
+    return new Promise(resolve => {
+        pool.query(
+            `SELECT id FROM sync_history WHERE user_id = ? ORDER BY sync_timestamp DESC LIMIT 1`,
+            [userId],
+            (err, res) => {
+                if (err) {
+                    console.log('select sync id', err);
+                    resolve(false);
+                } else {
+                    if (res.length) {
+                        resolve(res[0].id);
+                    } else {
+                        resolve(false);
+                    }
+                }
+            }
+        );
+    });
+}
+
 module.exports = {
-    getUserIdFromToken
+    getUserIdFromToken,
+    getSyncId,
+    getRecentSyncId
 };
